+++ AAEmu.Game/Models/Game/Skills/Effects/BuffEffect.cs
@@ -41,7 +41,7 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
             //if (target.Buffs.CheckBuffImmune(Buff.Id))
             //    return; // TODO send error of immune?
 
-            uint abLevel = 1;
+            short abLevel = 1;
             if (caster is Character character)
             {
                 if (source.Skill != null)
@@ -49,12 +49,12 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
                     var template = source.Skill.Template;
                     var abilityLevel = character.GetAbLevel((AbilityType)source.Skill.Template.AbilityId);
                     if (template.LevelStep != 0)
-                        abLevel = (uint)((abilityLevel / template.LevelStep) * template.LevelStep);
+                        abLevel = (short)((abilityLevel / template.LevelStep) * template.LevelStep);
                     else
-                        abLevel = (uint)template.AbilityLevel;
+                        abLevel = (short)template.AbilityLevel;
 
                     //Dont allow lower than minimum ablevel for skill or infinite debuffs can happen
-                    abLevel = (uint)Math.Max(template.AbilityLevel, (int)abLevel);
+                    abLevel = (short)Math.Max(template.AbilityLevel, (int)abLevel);
                 }
                 else if (source.Buff != null)
                 {
@@ -65,7 +65,7 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
             {
                 if(source.Skill != null)
                 {
-                    abLevel = (uint)source.Skill.Template.AbilityLevel;
+                    abLevel = (short)source.Skill.Template.AbilityLevel;
                 }
             }
